+++ package.json
@@ -16,6 +16,7 @@
     "react-cookie": "^7.2.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.26.2",
+    "sass": "^1.80.5",
     "swr": "^2.2.5"
   },
   "devDependencies": {
