+++ src/components/StoreMarker.module.scss
@@ -1,0 +1,9 @@
+.body {
+  color: black
+}
+
+.list {
+  list-style: none;
+  padding-left: 0;
+  display: table;
+}
